+++ spec/support/shared_contexts/with_mocked_kernel_warn.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.shared_context "with mocked `Kernel.warn` method" do
+  before { allow(Kernel).to receive(:warn).with(kind_of(String)) }
+end