+++ CHANGELOG.md
@@ -8,6 +8,7 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 Feature:
 - Drop Ruby MRI 2.7 support. (#194)
 - Add support for JRuby. (#193)
+- Emit a warning unless default value is a proc or an immutable object. (#191)
 - Allow `output` to accept procs as defaults. (#188)
 - Allow any object that responds to `===` to be used as a type. (#187)
 - Allow objects without `Kernel` inclusion to function with the `play` DSL. (#180)